[Console] Allow to pass the verbosity level to --verbose
The help text for the verbosity option mentions that it is possible to pass a value to set the verbosity level:

```
  -v|vv|vvv, --verbose  Increase the verbosity of messages: 1 for normal output, 2 for more verbose output and 3 for debug
```

However, when I try this I get an error message:

```
$ php application.php demo:greet Fabien --verbose=3

  [Symfony\Component\Console\Exception\RuntimeException]  
  The "--verbose" option does not accept a value.         
```

The `-vvv` option works fine, but the POSIX long option `--verbose` doesn't accept a value. This is clearly the intention since these values are being read in `Application::configureIO()`:

```
            if ($input->hasParameterOption('-vvv', true) || $input->hasParameterOption('--verbose=3', true) || $input->getParameterOption('--verbose', false, true) === 3) {
```

The solution is to simply change the input option from `VALUE_NONE` to `VALUE_OPTIONAL`.